Batch vertices for large shapes.
Fix for this bug: https://github.com/vydd/sketch/issues/160

The draw buffer isn't large enough for shapes that consist of
more than about 6000 vertices, so we need to batch the draw
calls. Different batching behaviour is required for different
primitive types, e.g. for triangle strips, in each batch you
have to include the last 2 vertices from the previous batch so
that the strip is continuous. I've only covered the :triangles
and :triangle-strip primitive types.