Use cffi:make-shareable-byte-vector to store canvas pixels.
This allows us to access and modify the array with AREF instead
of cffi:mem-aref, resulting in neater codes. It also avoids the
unnecessary cons-ing that happens when accessing arrays via the
CFFI interface.

The speed is about the same, maybe slightly faster; ~0.01 seconds
to write to all pixels in a 400x400 canvas. I compared this to the
speed of writing to all values in a byte array of size 4x400x400, and
the speed was 0.0002 seconds, 2 orders of magnitude faster. Maybe
it's the overhead of calling the PAINT function!